requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
tensorflow==2.8.0
pillow==9.0.1
requests==2.27.1
numpy==1.21.5
pydantic==1.9.0
python-dotenv==0.19.2
scikit-learn==1.0.2
joblib==1.1.0
